+++ lib/ui/page/home/page/chat/view.dart
@@ -259,7 +259,7 @@ class ChatView extends StatelessWidget {
                                               Flexible(
                                                 child: Obx(() {
                                                   return Text(
-                                                    c.chat!.title,
+                                                    c.chat!.title(),
                                                     style: style
                                                         .fonts
                                                         .big