+++ CHANGELOG.md
@@ -11,6 +11,12 @@ All user visible changes to this project will be documented in this file. This p
 
 [Diff](/../../compare/v0.6.6...v0.7.0) | [Milestone](/../../milestone/50)
 
+### Changed
+
+- UI:
+    - Home page:
+        - Display "Deleted Account" title for deleted users. ([#1419])
+
 ### Fixed
 
 - Web: